Conflicts:
	Inc/application/solverMaze/solverMaze.h

+++ Inc/peripherals/telemeters/telemeters.h
@@ -32,10 +32,10 @@
 /* Types definitions */
 typedef struct
 {
-    int16_t adc_value;
-    int16_t telemeter_values[SIZE_OF_AVEVAGE_TABLE];
-    int16_t value_average;
-    int64_t offset;
+    int32_t adc_value;
+    int32_t telemeter_values[SIZE_OF_AVEVAGE_TABLE];
+    int32_t value_average;
+    int32_t offset;
     uint16_t higher_interval;
     uint16_t lower_interval;
     char wall_presence;
@@ -53,8 +53,8 @@ typedef struct
 	telemeter_state left_diag;
 	telemeter_state ref_left_diag;
 
-	uint32_t it_cnt;
-	uint32_t end_of_conversion;
+	uint64_t it_cnt;
+	uint64_t end_of_conversion;
 	char active_state;
 	char emitter_state;
 	char selector;
